benchmarks: add transaction benchmark cli and update docs
